+++ include/compose.bash
@@ -20,9 +20,15 @@ compose-pull() {
 
     [ -f docker-compose.yml ] || deployer-generate
 
-    # parallell pull
-    # sed -n "s/.*image: //p" docker-compose.yml | sort -u | xargs -I@ -n1 bash -c "docker pull @ &"
     dockerCompose pull
+}
+
+compose-pull-parallel() {
+    declare desc="Pulls service images parallel"
+
+    [ -f docker-compose.yml ] || deployer-generate
+    
+    sed -n "s/.*image: //p" docker-compose.yml | xargs -n1 -P 20 docker pull
 }
 
 create-logfile() {